The tasklist menu
In this stage, we will implement a new menu that will prompt users to either enter a task or print all tasks. Since each task is added individually, it is possible to enter a multiple-line task.

Each task input should end when users input a blank line (or just spaces and tabs).