Date and time
In this stage, we will add on another useful feature — the ability to indicate a deadline for a particular task. We will also work on a priority tag feature.

Of course, a task may come without a deadline. However, to keep things simple, a deadline is required for each task in this stage.

When users add a task, also ask them to input a date, time, and a priority tag. The latter is just one letter: C, H, N, L (case-insensitive) per task that stands for Critical, High, Normal, and Low, respectively.