Edit the list
In this stage, we will implement the ability to delete or edit a task. In addition, we will also add a tag that signifies whether a task date is due or past.

The process of removing a task is based on the tasks sequence numbers. After deletion, the sequence numbers change.

Editing a task requires the change of priority, date, time, or task description. Due to IntelliJ run terminal limitations, editing can not be done over the data from the old task. That is, printing the old data and then editing them is not possible. So, for each of the fields above, you need to make another one and replace the old.

A due tag will point out whether a task is overdue or not. This is just one letter — I, T, or O (case-insensitive). They stand for In time, Today, and Overdue. To define the due tag, we will compare the current day with the task date.