Save the list
A task listing program is meaningless if each time you terminate it, the list of tasks is erased. We need a method of keeping tasks when the program isn't running. There are a good number of solutions: a database, several files, or even a cloud. In this stage, we are going to save the tasklist data locally into a file.